+++ src/Pages/Panel/Minimongo/Minimongo.tsx
@@ -8,7 +8,7 @@ import styled from 'styled-components'
 import { MinimongoStatus } from '@/Pages/Panel/Minimongo/MinimongoStatus'
 import { Button } from '@/Components/Button'
 import prettyBytes from 'pretty-bytes'
-import { MinimongoQueryDialog } from './MinimongoQueryDialog'
+import { MinimongoQueryDrawer } from './MinimongoQueryDrawer'
 
 interface Props {
   isVisible: boolean
@@ -116,7 +116,7 @@ export const Minimongo: FunctionComponent<Props> = observer(({ isVisible }) => {
       <MinimongoStatus />
 
       <MinimongoNavigator />
-      <MinimongoQueryDialog />
+      <MinimongoQueryDrawer />
     </Hideable>
   )
 })